Fix w64 build
* src/cryptcontroller.cpp (CryptController::resolve_keys): Use
SIZE_T_FORMAT and hwnd conversion.
* src/overlay.cpp (Overlay::Overlay): HWND cast to int is ok.

+++ src/overlay.cpp
@@ -46,7 +46,7 @@ Overlay::Overlay (HWND wid, const std::string &text): m_wid (wid)
   args.push_back (overlayer);
 
   args.push_back (std::string ("--hwnd"));
-  args.push_back (std::to_string ((int) wid));
+  args.push_back (std::to_string ((int) (intptr_t) wid));
 
   args.push_back (std::string ("--overlayText"));
   args.push_back (text);
